# services/api/tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Deployment": {
        "additionalProperties": false,
        "properties": {
          "base_token_address": {
            "title": "Base Token Address",
            "type": "string"
          },
          "chain_id": {
            "title": "Chain Id",
            "type": "integer"
          },
          "code_hashes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Code Hashes",
            "type": "object"
          },
          "compiler": {
            "additionalProperties": true,
            "title": "Compiler",
            "type": "object"
          },
          "deployed_at": {
            "title": "Deployed At",
            "type": "string"
          },
          "deployment_id": {
            "title": "Deployment Id",
            "type": "string"
          },
          "ens": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ens"
          },
          "exchange_address": {
            "title": "Exchange Address",
            "type": "string"
          },
          "explorer_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explorer Base Url"
          },
          "operator_address": {
            "title": "Operator Address",
            "type": "string"
          },
          "protocol_version": {
            "title": "Protocol Version",
            "type": "string"
          },
          "quote_token_address": {
            "title": "Quote Token Address",
            "type": "string"
          },
          "relay_address": {
            "title": "Relay Address",
            "type": "string"
          }
        },
        "required": [
          "deployment_id",
          "chain_id",
          "protocol_version",
          "exchange_address",
          "base_token_address",
          "quote_token_address",
          "operator_address",
          "relay_address",
          "code_hashes",
          "compiler",
          "explorer_base_url",
          "ens",
          "deployed_at"
        ],
        "title": "Deployment",
        "type": "object"
      },
      "DeploymentList": {
        "additionalProperties": false,
        "properties": {
          "deployments": {
            "items": {
              "$ref": "#/components/schemas/Deployment"
            },
            "title": "Deployments",
            "type": "array"
          }
        },
        "required": [
          "deployments"
        ],
        "title": "DeploymentList",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Scenario": {
        "additionalProperties": false,
        "description": "**Private** (both mandates): behind the observer reveal header.",
        "properties": {
          "buyer": {
            "additionalProperties": true,
            "title": "Buyer",
            "type": "object"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "public_config": {
            "additionalProperties": true,
            "title": "Public Config",
            "type": "object"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "seller": {
            "additionalProperties": true,
            "title": "Seller",
            "type": "object"
          }
        },
        "required": [
          "scenario_id",
          "name",
          "description",
          "public_config",
          "buyer",
          "seller"
        ],
        "title": "Scenario",
        "type": "object"
      },
      "ScenarioList": {
        "additionalProperties": false,
        "properties": {
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioSummary"
            },
            "title": "Scenarios",
            "type": "array"
          }
        },
        "required": [
          "scenarios"
        ],
        "title": "ScenarioList",
        "type": "object"
      },
      "ScenarioSummary": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "feasibility_hint": {
            "title": "Feasibility Hint",
            "type": "null"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "scenario_id",
          "name",
          "description",
          "feasibility_hint"
        ],
        "title": "ScenarioSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The operator API of docs/api_contract.md section 2. Test assets and simulated economics only.",
    "title": "Agent Negotiation Sandbox \u2014 operator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/deployments": {
      "get": {
        "operationId": "list_deployments_v1_deployments_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Deployments",
        "tags": [
          "system"
        ]
      }
    },
...
    "/v1/scenarios": {
      "get": {
        "description": "The public fields only. `feasibility_hint` is always null: feasibility is computed by the\nevaluator alone, never on a setup route.",
        "operationId": "list_scenarios_v1_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "system"
        ]
      }
    },
    "/v1/scenarios/{scenario_id}": {
      "get": {
        "description": "The whole template, both mandates included, so behind the observer reveal header.",
        "operationId": "get_scenario_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Scenario",
        "tags": [
          "system"
        ]
      }
    }
  }
}